Move sliders' *.ts files under widgets/sliders/
