+++ bokehjs/src/lib/models/widgets/sliders/categorical_slider.ts
