+++ bokehjs/src/lib/models/widgets/sliders/index.ts
@@ -1,6 +1,6 @@
-export {CategoricalSlider}   from "../categorical_slider"
-export {DateRangeSlider}     from "../date_range_slider"
-export {DateSlider}          from "../date_slider"
-export {DatetimeRangeSlider} from "../datetime_range_slider"
-export {RangeSlider}         from "../range_slider"
-export {Slider}              from "../slider"
+export {CategoricalSlider}   from "./categorical_slider"
+export {DateRangeSlider}     from "./date_range_slider"
+export {DateSlider}          from "./date_slider"
+export {DatetimeRangeSlider} from "./datetime_range_slider"
+export {RangeSlider}         from "./range_slider"
+export {Slider}              from "./slider"
