+++ bokehjs/src/lib/models/widgets/sliders/numerical_range_slider.ts
