+++ bokehjs/src/lib/models/widgets/sliders/numerical_slider.ts
